# tests/contract/openapi.yaml
openapi: 3.0.3
info:
  title: SDET Automation Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /cart:
    get:
      summary: Get cart
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CartItem" }
...
  /checkout:
    post:
      summary: Place order
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coupon: { type: string }
                card_token: { type: string }
      responses:
        "201": { description: Created }
        "200": { description: Idempotent replay }
        "400": { description: Bad request }
        "402": { description: Payment failed }
        "409": { description: Insufficient stock }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CartItem:
      type: object
      required: [sku, qty, name, price_cents]
      properties:
        sku: { type: string }
        qty: { type: integer, minimum: 1 }
        name: { type: string }
        price_cents: { type: integer, minimum: 0 }